+++ pyobjc-framework-CoreAudio/Modules/_CoreAudio_AudioBuffer.m
@@ -5,9 +5,9 @@
  * basically a buffer with extra attributes.
  */
 
-static PyTypeObject audio_buffer_type; /* Forward definition */
+static PyObject* audio_buffer_type;
 
-#define audio_buffer_check(obj) PyObject_TypeCheck(obj, &audio_buffer_type)
+#define audio_buffer_check(obj) PyObject_TypeCheck(obj, (PyTypeObject*)audio_buffer_type)
 
 struct audio_buffer {
     PyObject_HEAD
@@ -157,7 +157,7 @@ ab_new(PyTypeObject* cls, PyObject* args, PyObject* kwds)
         return NULL;
     }
 
-    result = PyObject_New(struct audio_buffer, &audio_buffer_type);
+    result = PyObject_New(struct audio_buffer, (PyTypeObject*)audio_buffer_type);
     if (result == NULL) {
         return NULL;
     }
@@ -207,18 +207,24 @@ PyDoc_STRVAR(ab_doc, "AudioBuffer(*, num_channels=1, buffer_size=-1)\n" CLINIC_S
                      "this will allocate a buffer, otherwise no buffer is "
                      "allocated\n");
 
-static PyTypeObject audio_buffer_type = {
-    PyVarObject_HEAD_INIT(&PyType_Type, 0).tp_name = "CoreAudio.AudioBuffer",
-    .tp_basicsize                                  = sizeof(struct audio_buffer),
-    .tp_itemsize                                   = 0,
-    .tp_dealloc                                    = ab_dealloc,
-    .tp_getattro                                   = PyObject_GenericGetAttr,
-    .tp_flags                                      = Py_TPFLAGS_DEFAULT,
-    .tp_doc                                        = ab_doc,
-    .tp_methods                                    = ab_methods,
-    .tp_getset                                     = ab_getset,
-    .tp_members                                    = ab_members,
-    .tp_new                                        = ab_new,
+static PyType_Slot ab_slots[] = {
+    {.slot = Py_tp_dealloc, .pfunc = (void*)&ab_dealloc},
+    {.slot = Py_tp_getattro, .pfunc = (void*)&PyObject_GenericGetAttr},
+    {.slot = Py_tp_doc, .pfunc = (void*)&ab_doc},
+    {.slot = Py_tp_methods, .pfunc = (void*)&ab_methods},
+    {.slot = Py_tp_getset, .pfunc = (void*)&ab_getset},
+    {.slot = Py_tp_members, .pfunc = (void*)&ab_members},
+    {.slot = Py_tp_new, .pfunc = (void*)&ab_new},
+
+    {0, NULL} /* sentinel */
+};
+
+static PyType_Spec ab_spec = {
+    .name      = "CoreAudio.AudioBuffer",
+    .basicsize = sizeof(struct audio_buffer),
+    .itemsize  = 0,
+    .flags     = Py_TPFLAGS_DEFAULT | Py_TPFLAGS_HEAPTYPE,
+    .slots     = ab_slots,
 };
 
 static PyObject*
@@ -226,7 +232,7 @@ ab_create(AudioBuffer* item)
 {
     struct audio_buffer* result;
 
-    result = PyObject_New(struct audio_buffer, &audio_buffer_type);
+    result = PyObject_New(struct audio_buffer, (PyTypeObject*)audio_buffer_type);
     if (result == NULL) {
         return NULL;
     }
@@ -275,20 +281,30 @@ init_audio_buffer(PyObject* module)
 {
     int r;
 
-    if (PyType_Ready(&audio_buffer_type) == -1)
-        return -1;
-
-    r = PyDict_SetItemString(audio_buffer_type.tp_dict, "__typestr__",
-                             PyBytes_FromString(@encode(AudioBuffer)));
-    if (r == -1)
-        return -1;
-
-    Py_INCREF(&audio_buffer_type);
-    r = PyModule_AddObject(module, "AudioBuffer", (PyObject*)&audio_buffer_type);
-    if (r == -1) {
-        Py_DECREF(&audio_buffer_type);
+    PyObject* tmp = PyType_FromSpec(&ab_spec);
+    if (tmp == NULL) {
         return -1;
     }
+    audio_buffer_type = tmp;
+
+    PyObject* ts = PyBytes_FromString(@encode(AudioBuffer));
+    if (ts == NULL) {
+        Py_CLEAR(audio_buffer_type);
+        return -1;
+    }
+    r = PyObject_SetAttrString(audio_buffer_type, "__typestr__", ts);
+    Py_DECREF(ts);
+    if (r == -1) {
+        Py_CLEAR(audio_buffer_type);
+        return -1;
+    }
+
+    r = PyModule_AddObject(module, "AudioBuffer", audio_buffer_type);
+    if (r == -1) {
+        Py_CLEAR(audio_buffer_type);
+        return -1;
+    }
+    Py_INCREF(audio_buffer_type);
 
     r = PyObjCPointerWrapper_Register("AudioBuffer*", @encode(AudioBuffer*),
                                       pythonify_audio_buffer, depythonify_audio_buffer);